// sm_vue_client/src/openapi/system/models/VoloAbpIdentityIdentityRoleDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VoloAbpIdentityIdentityRoleDto = {
    readonly extraProperties?: Record<string, any> | null;
    id?: string;
    name?: string | null;
    isDefault?: boolean;
    isStatic?: boolean;
    isPublic?: boolean;
    concurrencyStamp?: string | null;
};