// sm_vue_client/src/openapi/system/models/VoloAbpIdentityIdentityRoleUpdateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VoloAbpIdentityIdentityRoleUpdateDto = {
    readonly extraProperties?: Record<string, any> | null;
    name: string;
    isDefault?: boolean;
    isPublic?: boolean;
    concurrencyStamp?: string | null;
};